+++ uni_modules/uview-ui/components/u-picker/props.js
@@ -74,6 +74,11 @@ export default {
         defaultIndex: {
             type: Array,
             default: uni.$u.props.picker.defaultIndex
-        }
+        },
+		// 是否在手指松开时立即触发 change 事件。若不开启则会在滚动动画结束后触发 change 事件，只在微信2.21.1及以上有效
+		immediateChange: {
+			type: Boolean,
+			default: uni.$u.props.picker.immediateChange
+		}
     }
 }